Use split factors directly instead of idents
I guess this would be considered constant folding? The reason for
this change is to support storage folding where the split factors
need to be defined in the executive function and in the kernel
functions. Rather than redefine (which is hacky) or pass as args
(which is a headache), we'll just use the constants since they're
known at compile time. The downside to this is worse debuggability.