[ChatGPT-o3] Rewrite `Graph` to just hold roots
This was my prompt: "here is graph.rs from my deep learning compiler. right now
Graph has a nodes field but this is not really necessary. instead, it should
have a roots field. specifically, it should be possible for Graphs to have more
than one root. can you make this change for me?"