Add lowerer cache to properly handle DAGs
Previously the lowerer would relower nodes already visited via
another path from the root node leading to extraneous inputs in the
generated code. Now it caches `Lowered` results of interior nodes
and guards leaf nodes from being relowered. For the interior nodes,
we strip out the def_block and the exec_block so they are not
duplicated. This feels like a hack, and this should probably be
rearchitected eventually.