Fix overshadowing of reconstructed index
Generate a declaration of a temp variable so that the loop index
is not overshadowed by the reconstructed index (on tiled loops).
This is does because in C, the loop index variable is immediately
overshadowed in the declaration and cannot be used to update the
value of the loop index. For example, in

```c
size_t i1 = (((b1_0_1)*i1) + i1_0);
```

`i1` on the rhs is indeterminate because of the declaration of new
of `i1` on the lhs. This cause `i1` to be set to an incorrect value.

Instead, we can put the value into a temp variable and then use it
in the reconstruction assignment, like this:

```c
size_t tmp = (((b1_0_1)*i1) + i1_0);
size_t i1 = tmp;`
```